Work towards supporting 16-bit values for the storage-area
This pull request, once complete, will support 16-bit values being
input and output from the storage-area.

Currently I've reworked the output to show the contents of the area
as 4-byte hex values.  The next commit will parse input values as
1, 2, 3, or 4, byte hex inputs.

So we'll expect to allow the user to enter each of the following
statements:

* 3#
* 3F#
* 322#
* ffFF#

Once complete this will close #1.