Allow reading/writing/JMP to RAM
This pull-request introduces the idea of a memory-variable, which
can be used to either read, write, or JMP to an area of memory.

For example to dump the contents of 0x0000:

c  ; clear temporary storage area
m  ; set address to be 0x0000
rp ; read and print contents
rp ; read and print ..
..

c m g ; jmp to 0x0000

This closes #2.